iniziato refactoring del codice condiviso
