+++ es2/Marco Ottina 795130 - es 2.py
@@ -6,10 +6,16 @@ import re
 import csv
 import collections
 
+import sys
+sys.path.append(".\\..\\..\\utils")
+import utils
+import functions
+import cacheVarie
 
-regex_punctuation = r'[^\w\s]'
-lemmatizer = WordNetLemmatizer()
-stop_words = set(stopwords.words('english'))
+
+#regex_punctuation = r'[^\w\s]'
+#lemmatizer = WordNetLemmatizer()
+#stop_words = set(stopwords.words('english'))
 
 
 
@@ -19,126 +25,6 @@ stop_words = set(stopwords.words('english'))
 #------------
 #------------
 
-
-
-#------------ start imported
-
-
-def jaccard_similarity(sent1, sent2):
-	len_sent1 = len(sent1)
-	len_sent2 = len(sent2)
-	if len_sent1== 0 or len_sent2 == 0:
-		return 0
-	intCard = len(sent1.intersection(sent2))
-	union_card = (len_sent1 + len_sent2) - (intCard)
-	return intCard/union_card
-	
-def overlap_similarity(sent1, sent2):
-	len_sent1 = len(sent1)
-	len_sent2 = len(sent2)
-	if len_sent1 == 0 or len_sent2 == 0:
-		if len_sent1 == 0 and len_sent2 == 0:
-			return 1
-		else:
-			return 0
-	intCard = len(sent1.intersection(sent2))
-	return intCard / min(len_sent1, len_sent2)
-
-def compute_sim(sent1, sent2, usingOverlapSimilarity = True):
-	if usingOverlapSimilarity:
-		return overlap_similarity(sent1, sent2)
-	else:
-		return jaccard_similarity(sent1, sent2)
-
-
-#------------ end imported
-
-
-
-def print_tab(level, end_line = "\n"):
-	if level > 0:
-		print("\t" * level, end = end_line)
-
-def deep_array_printer(x, level = 0):
-	if isinstance(x, list):
-		print('[')
-		new_level = level + 1
-		i = 0
-		for y in x:
-			print_tab(new_level, "")
-			print(i, '\t -> ', end = "")
-			deep_array_printer(y, new_level)
-			i += 1
-		print_tab(level, "")
-		print(']')
-	else:
-		print(x)
-
-
-def clear_cache():
-	cache_synsets = {}
-
-
-def split_row_string(rs):
-	start = 0
-	length = len(rs)
-	parts = []
-	if rs[0] == '\"':
-		i = 1
-		start = 1
-		delimiter = '\"'
-	else:
-		i = 0
-		delimiter = ','
-	while i < length:
-		if rs[i] is delimiter:
-			if i != start:
-				parts.append(rs[start:i])
-			else: # record the empty string
-				parts.append("")
-			#move to the next chunk
-			if delimiter == '\"':
-				i += 2
-			else:
-				i += 1
-			start = i
-			if start < length:
-				if rs[start] == '\"':
-					delimiter = '\"'
-					i += 1
-					start = i
-				else:
-					delimiter = ','
-			else:
-				if rs[length-1] == ',':
-					parts.append("")
-		else:
-			i += 1
-	if i > length:
-		i = length
-	if start < i: # add the last part that I forgot
-		parts.append(rs[start:i] )
-	return parts
-
-
-def filter_and_lemmatize_words_in(text):
-	no_punct = re.sub(regex_punctuation, '', text.lower())
-	words = no_punct.split()
-	words = list(filter(lambda x: x not in stop_words, words))
-	return set([lemmatizer.lemmatize(w) for w in words])
-	
-def synsetToBagOfWords(s, useExamples = True):
-	bag = set()
-	texts_to_process = collections.deque([l.name() for l in s.lemmas()])
-	texts_to_process.append(s.definition())
-	if useExamples:
-		for e in s.examples():
-			texts_to_process.append(e)
-	for t in texts_to_process:
-		lemm = filter_and_lemmatize_words_in(t)
-		for l in lemm:
-			bag.add(l)
-	return bag
 
 
 def read_csv():
@@ -152,7 +38,7 @@ def read_csv():
 				if can_use:
 					row_string = row[0]
 					i = 0
-					cols_in_this_row = split_row_string(row_string)
+					cols_in_this_row = split_csv_row_string(row_string)
 					length = len(cols_in_this_row) -1
 					while i < length:
 						cols[i].append(cols_in_this_row[i + 1]) #because the first column is reserved for indexes
@@ -189,26 +75,7 @@ print(bag_root_synset)
 
 
 
-
-
-def get_synsets(word_text, cache_synsets_by_name = None):
-	if cache_synsets_by_name is None:
-		cache_synsets_by_name = cache_synsets
-	if word_text in cache_synsets:
-		return cache_synsets_by_name[word_text]
-	s = wn.synsets(word_text)
-	cache_synsets_by_name[word_text] = s
-	return s
-
-def get_synset_bag(syns, cache_synsets_bag_by_name = None):
-	if cache_synsets_bag_by_name is None:
-		cache_synsets_bag_by_name = cache_synsets_bag
-	name = syns.name()
-	if name in cache_synsets_bag_by_name:
-		return cache_synsets_bag_by_name[name]
-	b = synsetToBagOfWords(syns)
-	cache_synsets_bag_by_name[name] = b
-	return b
+
